moved interest to exchange class method
