+++ freqtrade/exchange/kraken.py
@@ -1,6 +1,8 @@
 """ Kraken exchange subclass """
 import logging
 from datetime import datetime
+from decimal import Decimal
+from math import ceil
 from typing import Any, Dict, List, Optional, Tuple
 
 import ccxt
@@ -206,3 +208,26 @@ class Kraken(Exchange):
             fees = sum(df['open_fund'] * df['open_mark'] * amount * time_in_ratio)
 
         return fees if is_short else -fees
+
+    @classmethod
+    def interest(
+        cls,
+        borrowed: Decimal,
+        rate: Decimal,
+        hours: Decimal
+    ) -> Decimal:
+        """
+        Equation to calculate interest on margin trades
+
+        :param borrowed: The amount of currency being borrowed
+        :param rate: The rate of interest (i.e daily interest rate)
+        :param hours: The time in hours that the currency has been borrowed for
+
+        Returns: The amount of interest owed (currency matches borrowed)
+        """
+
+        one = Decimal(1.0)
+        four = Decimal(4.0)
+
+        # Rounded based on https://kraken-fees-calculator.github.io/
+        return borrowed * rate * (one+ceil(hours/four))
