--- freqtrade/leverage/__init__.py
@@ -1,2 +1,0 @@
-# flake8: noqa: F401
-from freqtrade.leverage.interest import interest
